requirements.txt
dash>=2.1.0
dash-bootstrap-components>=1.1.0
dash-holoniq-components>=0.0.19
iniconfig>=1.1.1

# setup.py - exclude

# For ticker demo

pandas>=1.4.2
colorlover>=0.3.0

# For admin pages demo support

Flask_Login>=0.4.1
Flask_SQLAlchemy>=2.5.1
SQLAlchemy_Utils>=0.38.2
cachetools>=5.0.0
pystache>=0.6.0
dash-datatables>=0.0.9

# Waitress server

waitress>=2.1.1
Paste>=3.5.0


# For package development & testing

pytest>=7.1.1
setuptools==41.0.1
twine==3.7.1
keyrings.alt>=4.1.0
selenium>=4.1.0
dash[testing]>=2.1.0
pytest-env>=0.6.2
pytest-mock>=3.7.0

